+++ BusinessEntity/Program.cs
@@ -14,6 +14,7 @@ using BusinessEntity.MiniApps.UserMiniApp.Storage;
 using BusinessEntity.Service;
 using BusinessEntity.Service.WebLogging;
 using BusinessEntity.Services;
+using BusinessEntity.Services.Backup;
 using BusinessEntity.Services.RichTextImport;
 using BusinessEntity.Settings;
 using BusinessEntity.WebLogger.Services;
@@ -142,6 +143,9 @@ namespace BusinessEntity
             builder.Services.AddScoped<BusinessEntity.Contracts.IUserContextService, BusinessEntity.Services.UserContextService>();
             builder.Services.AddScoped<ReactiveUI.IMessageBus, ReactiveUI.MessageBus>();
             builder.Services.AddScoped<BusinessEntity.Services.ITreeSelectionService, BusinessEntity.Services.TreeSelectionService>();
+            builder.Services.AddSingleton<IBusinessEntityBackupHandler, GenericBusinessEntityBackupHandler>();
+            builder.Services.AddSingleton<SpaceBackupService>();
+            builder.Services.AddHostedService(provider => provider.GetRequiredService<SpaceBackupService>());
 
             // Настраивает EF Core и фабрику DbContext для работы с БД.
             var connectionString = builder.Configuration.GetConnectionString("DockerConnection");